+++ src/modes/cutscene_world.hpp
@@ -42,14 +42,14 @@ class CutsceneWorld : public World
     std::map<float, std::vector<TrackObject*> > m_sounds_to_stop;
     std::map<float, std::vector<TrackObject*> > m_particles_to_trigger;
     
-    float m_duration;
+    double m_duration;
     bool m_aborted;
     
     /** monkey tricks to get the animations in sync with irrlicht. we reset the time
      *  after all is loaded and it's running withotu delays
      */
     bool m_second_reset;
-    float m_time_at_second_reset;
+    double m_time_at_second_reset;
     
     void abortCutscene()
     {
@@ -85,7 +85,7 @@ public:
     virtual const std::string& getIdent() const OVERRIDE;
     
     virtual void kartHit(const int kart_id) OVERRIDE;
-    virtual void update(float dt) OVERRIDE;
+    virtual void update(double dt) OVERRIDE;
 
     virtual void createRaceGUI() OVERRIDE;
